.history/requirements_20240201210233.txt
pandas
mlflow==2.2.2
notebook
numpy
scikit-learn
matplotlib
python-box==6.0.2
pyYAML
tqdm
ensure==1.0.2
joblib
types_PyYAML
Flask
Flask-Cors
-e .
